pyproject.toml
[project]
name = "naoteam_boost_uc_ball_hyp_generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "ultralytics>=8.3.184",
    "torchprofile>=0.0.4",
    "matplotlib>=3.0.0",
    "pillow>=8.0.0",
    "numpy>=1.21.0",
    "tensorboard>=2.0.0",
    "torch2trt>=0.4.0",
    "tensorrt>=8.0.0",
    "pycuda>=2021.1",
]

[dependency-groups]
dev = ["ruff>=0.12.10", "ty>=0.0.1a19"]

[tool.ruff]
line-length = 120

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]
